backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# Date/time
python-dateutil==2.8.2